multi-agent-ai-app/requirements.txt
openai>=1.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
language-tool-python>=2.7.0